Work_Version

+++ Assets/Source/Codebase/Domain/Configs/ItemConfig.cs
@@ -6,7 +6,7 @@ namespace Source.Codebase.Domain.Configs
     public class ItemConfig : ScriptableObject
     {
         [field: SerializeField] public string Name { get; private set; }
-        [field: SerializeField] public float Price { get; private set; }
+        [field: SerializeField] public int Price { get; private set; }
         [field: SerializeField] public int ClickForce { get; private set; }
         [field: SerializeField] public Sprite Icon { get; private set; }
         [field: SerializeField] public ClickType ClickType { get; private set; }
